// APPADMON001SM/APPADMON001MW/src/app/modules/login/models/clsLoginModels.ts
export class clsUsuarioMenu {
    idEmpleado: number = 0;
    usuario: string = '';
    idCargo: number = 0;
    idTipoUsuario: number = 0;
    idEncabezado: number = 0;
    idOpcionEnc: number = 0;
    posEnc: number = 0;
    tituloEncabezado: string = '';
    rutaEncabezado: string = '';
    icono: string = '';
    idOpcionDet: number = 0;
    posDet: number = 0;
    tituloDetalle: string = '';
    SubRuta: string = '';
    linkDet: string = '';
}